First cut at new decoder and uop cracker.
Major changes:
- getting rid of the tables in machine.* and x86flow.*
- .. and a lot more (now) dead code
- using XED for decoding instead
- getting rid of all kinds of actual values
- getting rid of oracle speculative memory
- removed (most of) special-case rflags handling

Minor:
- streamlined fp_penalty
- got rid of several special allocation pools

Many things TODO:
- get rid of register values all the way from the feeder
- decoder features: immediates, REP microbranches
- decoder: microcoded instructions (push, pop, etc.)
- uop-cracker: actual mapping to FUs
- speculation
- oracle still isn't fully working (especially wrt speculation)

+++ ZPIPE-commit/commit-STM.cpp
@@ -214,7 +214,7 @@ void core_commit_STM_t::step(void)
         struct uop_t * uop = &Mop->uop[Mop->commit.complete_index];
 
         Mop->commit.complete_index += uop->decode.has_imm ? 3 : 1;
-        if(Mop->commit.complete_index >= Mop->decode.flow_length)
+        if(Mop->commit.complete_index >= (int) Mop->decode.flow_length)
         {
           Mop->commit.complete_index = -1; /* Mark this Mop as all done */
           if(Mop->fetch.bpred_update)
@@ -273,7 +273,7 @@ void core_commit_STM_t::step(void)
       /* mark uop as committed in Mop */
       Mop->commit.commit_index += uop->decode.has_imm ? 3 : 1;
 
-      if(Mop->commit.commit_index >= Mop->decode.flow_length)
+      if(Mop->commit.commit_index >= (int) Mop->decode.flow_length)
       {
         Mop->commit.commit_index = -1; /* The entire Mop has been committed */
